Connect using system selected local port
Rather than using the port number from the camera end
pass the symbolic value 0 in as the port number. This allows
the system to choose any available port number to connect on.
The socket can then be queried to get the actual port number.

This change allows multiple cameras to be active at the same time
as they are now using system selected free local port numbers.